requirement.txt
requests
sqlalchemy
pandas
psycopg2
python-dotenv==0.19.1
